fixed visualization, still need to do a bunch of refactoring before moving on
