+++ Tests/example2/server.py
@@ -5,7 +5,7 @@ import Nodes.initializers as initializers
 import os
 
 # GRAPH CREATION
-G = nx.erdos_renyi_graph(10, 0.5, seed=123, directed=False)
+G = nx.erdos_renyi_graph(20, 0.5, seed=1, directed=False)
 n = G.number_of_nodes()
 m = G.number_of_edges()
 #utils.draw_graph(G)
@@ -16,9 +16,10 @@ print(f"With flood algorithm, we expect {(2*m)-(n-1)} messages.")
 # FRAMEWORK
 client = os.path.abspath("./client.py")
 print(client)
-init = initializers.Initializer(client, "localhost", 65000, G, shell=False)
+init = initializers.Initializer(client, "localhost", 65000, G, shell=False, visualizer=True)
 init.initialize_clients()
 init.setup_clients()
 init.wakeup(5)
 #init.wakeup_all(2)
+init.start_visualization()
 init.wait_for_number_of_messages()
